+++ packages/core/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @scalar/core
 
+## 0.3.3
+
+### Patch Changes
+
+- Updated dependencies [221e35f]
+  - @scalar/types@0.2.3
+
+## 0.3.2
+
+### Patch Changes
+
+- @scalar/types@0.2.2
+
 ## 0.3.1
 
 ### Patch Changes