+++ packages/openapi-parser/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @scalar/openapi-parser
 
+## 0.17.0
+
+### Minor Changes
+
+- f823d45: feat(openapi-parser): introduce parseJson and parseYaml plugins for bundler
+
+### Patch Changes
+
+- 166e298: feat(openapi-parser): correctly set the base origin for string inputs for the bundler
+- 4156f1d: Expand workspace store integration
+- 37c90b8: feat: add x-webhooks upgrade to the upgrader
+
+## 0.16.0
+
+### Minor Changes
+
+- 11fabae: fix(openapi-parser): webpack bundle errors because of the barrel files
+
+## 0.15.0
+
+### Minor Changes
+
+- b9dacba: fix(openapi-parser): multi entry build for bundler plugins
+- a73e9be: chore(openapi-parser): bring back pipeline syntax
+- f4332eb: feat: external reference resolution and partial bundle of the openapi document
+
+### Patch Changes
+
+- 17bc0d8: fix: collectionFormat is not migrated to new style and explode keywords
+- 49c04cf: fix(openapi-parser): use dynamic imports inside the plugin
+
 ## 0.14.0
 
 ### Minor Changes