+++ packages/workspace-store/CHANGELOG.md
@@ -1,5 +1,69 @@
 # @scalar/workspace-store
 
+## 0.4.1
+
+### Patch Changes
+
+- Updated dependencies [7913193]
+  - @scalar/code-highlight@0.1.4
+
+## 0.4.0
+
+### Minor Changes
+
+- 7338d56: feat(workspace-store): generate sidebar entries
+
+### Patch Changes
+
+- 4156f1d: Expand workspace store integration
+- Updated dependencies [221e35f]
+- Updated dependencies [166e298]
+- Updated dependencies [4156f1d]
+- Updated dependencies [a37df33]
+- Updated dependencies [f823d45]
+- Updated dependencies [37c90b8]
+  - @scalar/openapi-types@0.3.3
+  - @scalar/helpers@0.0.4
+  - @scalar/openapi-parser@0.17.0
+  - @scalar/code-highlight@0.1.3
+
+## 0.3.2
+
+### Patch Changes
+
+- 11fabae: fix(openapi-parser): webpack bundle errors because of the barrel files
+- Updated dependencies [11fabae]
+  - @scalar/openapi-parser@0.16.0
+
+## 0.3.1
+
+### Patch Changes
+
+- 373214a: fix(workspace-store): remove readFiles plugin from the client store
+
+## 0.3.0
+
+### Minor Changes
+
+- bbad7c7: feat(workspace-store): create a sync client store
+
+## 0.2.0
+
+### Minor Changes
+
+- f4332eb: feat: external reference resolution and partial bundle of the openapi document
+
+### Patch Changes
+
+- Updated dependencies [17bc0d8]
+- Updated dependencies [b9dacba]
+- Updated dependencies [a73e9be]
+- Updated dependencies [49c04cf]
+- Updated dependencies [f4332eb]
+- Updated dependencies [05c22c7]
+  - @scalar/openapi-parser@0.15.0
+  - @scalar/openapi-types@0.3.2
+
 ## 0.1.1
 
 ### Patch Changes